+++ src/ttc_control.c
@@ -13,12 +13,12 @@
  * @return The acceleration in m/s². If the function is called for the first time or there is
  *         insufficient time difference between calls, it returns 0.0.
  */
-float accel_calc(float spd) {
-    static float prev_spd = 0.0;
+double accel_calc(double spd) {
+    static double prev_spd = 0.0;
     static struct timespec start_time = {0, 0};
     struct timespec current_time;
     double elapsed_time;
-    float accel = 0.0;
+    double accel = 0.0;
 
     clock_gettime(CLOCK_REALTIME, &current_time);
 
@@ -56,8 +56,8 @@ float accel_calc(float spd) {
  *         the function returns -1.0. If there is no relative acceleration, the time to collision is calculated
  *         as distance divided by speed.
  */
-float ttc_calc(float dis_rel, float spd_rel) {
-    float a, b, c, ttc, delta;
+double ttc_calc(double dis_rel, double spd_rel) {
+    double a, b, c, ttc, delta;
     
     a = accel_calc(spd_rel);
     b = spd_rel / 3.6;
